Assignments

Documentation

data Configuration

A an assignment configuration data structure | Uses Data.Time.UTCTime | If files is an empty list, ANY number of files are OK

Constructors

Configuration UTCTime UTCTime UTCTime [String] Double Double Double

data Assignment

An assignment descriptor

Constructors

Assignment Year Type Int

type UserIdentifier = String

A user identifier (not DB id) like a username or JMBAG

type Year = Integer

Academic year shorthand (e.g. 2015 for 2015/16)

data Submission

Constructors

Submission Assignment UserIdentifier [FilePath]

data Type

An assignment type

Constructors

Homework

Exam

Project

getConfiguration :: Assignment -> IO Configuration

listSubmissions :: Assignment -> IO [UserIdentifier]

getSubmission :: Assignment -> UserIdentifier -> IO Submission

listFiles :: Submission -> IO [FilePath]

createAssignment :: Assignment -> Configuration -> FilePath -> IO ()

Creates a new assignment from Assignment, configuration and PDF file | The PDF file should be copied, moved or symlinked so that it is | accessible from the assignment directory.

getSubmissionPath :: Submission -> FilePath

upload :: Submission -> Text -> String -> IO Submission