Email

Documentation

type Template = Text

An alias for the template contents

data Configuration

Configuration object

compileTemplate :: Template -> Map String String -> Either Error Text

Parses an expression using traslateTags

getConfiguration :: FilePath -> IO Configuration

Get configuration from a file path

sendMail :: IO Configuration -> IO Text -> [User] -> String -> IO ()

Sends an e-mail with given text to list of users | using given configuration. Throws appropriate error upon failure. | Provide subject of the email as the last argument.