Reviews

Documentation

data ReviewAssignment

A review assignment representation

Constructors

ReviewAssignment UserIdentifier UserIdentifier Role Assignment

data Review

A finished review

Constructors

Review ReviewAssignment Double Text

data Role

A user s role or authorization level as a reviewer

Constructors

Student

Staff

assignNReviews :: Assignment -> [UserIdentifier] -> [UserIdentifier] -> Int -> Role -> IO [ReviewAssignment]

Takes an Assignment, a list of reviewer identifiers and a | list of reviewee identifiers and assigns N reviewees for each | reviewer. It makes sure that a user never reviews themselves. | The reviewer is assigned the reviews with the provided role.

assignReviews :: Assignment -> [UserIdentifier] -> [UserIdentifier] -> Role -> IO [ReviewAssignment]

storeAssigments :: [ReviewAssignment] -> IO ()

Stores a list of review assignments into a database or | file system.

assignedReviews :: Assignment -> IO [ReviewAssignment]

Retrieves all ReviewAssignments for an Assignment from | a database or file system.

assignmentsBy :: Assignment -> UserIdentifier -> IO [ReviewAssignment]

Retrieves all ReviewAssignments for an Assignment and | a UserIdentifier, i.e. all the reviews for that assigment | the user has to perform.

assignmentsFor :: Assignment -> UserIdentifier -> IO [ReviewAssignment]

Retrieves all ReviewAssignments for an Assignment and | a UserIdentifier, i.e. all the reviews for that assignment | where the user is a reviewee.

saveReview :: Review -> IO ()

Completes a review assignment and stores the result in a | file system or database.

reviews :: Assignment -> IO [Review]

Loads all the completed review results for an assignment

reviewsBy :: Assignment -> UserIdentifier -> IO [Review]

Loads all the completed review results for an assignment | that were performed by a user.

reviewsFor :: Assignment -> UserIdentifier -> IO [Review]

Loads all the completed review results for an assignment | that were performed by a user.