User

Documentation

createUser :: UserIdentifier -> String -> String -> Role -> IO User

Takes a user identifier, e-mail, password and role. | Performs password hashing and stores the user into the | database, returning a filled User. If creating it fails (e.g. | the user identifier is already taken), throws an appropriate | exception.

updateUser :: User -> IO ()

Updates a given user. Identifies it by the UserIdentifier (or | maybe database id field, if you added it) in the User and overwrites | the DB entry with the values in the User structure. Throws an | appropriate error if it cannot do so; e.g. the user does not exist.

deleteUser :: UserIdentifier -> IO ()

Deletes a user referenced by identifier. If no such user or the | operation fails, an appropriate exception is thrown.

listUsers :: IO [User]

Lists all the users

listUsersInRole :: Role -> IO [User]

Lists all users in a given role

getUser :: UserIdentifier -> IO User

Fetches a single user by identifier

isRoleInYear :: User -> Role -> Integer -> Bool

Checks whether the user has a role of AT LEAST X in a given academic | year.

type UserIdentifier = String